Only use racer for completions in rust-mode.
The default value for `completion-at-point-functions` is
`'(tags-completion-at-point-function)`. As a result, if the user has any
tags tables in `tagsl-table-list` then `company-capf` tries to use those
tags tables for completions.

Instead, we ensure that `completion-at-point-functions` is initialised
to nil, so rust-mode buffers don't end up using these unwanted other
completion sources.