Dynamically expand RUST_SRC_PATH and CARGO_HOME.
This commit changes the behavior of `racer--call` to dynamically expand
`RUST_SRC_PATH` and `CARGO_HOME`. This is useful if these environment variables
aren't set at the time emacs is started.

I'm working on a tool to dynamically apply development environments from
direnv[1] when opening files in Emacs. This is useful to set
`RUST_SRC_PATH` to different values for each project (without resorting
to directory-local variables). This commit is necessary to support this
in racer.

[1]: https://github.com/direnv/direnv