let `racer` locate rust sources by default
Racer looks for Rust sources by calling `rustc --print sysroot`, if
`$RUST_SRC_PATH` is not set. This plays well with rustup's capability
to manage rust sources along with the rest of the toolchain.

To make use of this heuristic, we no longer require a source path to be
explicitly configured by the user.