Do not expand environment variables
Only expand `racer-rust-src-path` and `racer-cargo-home` variables, not
`RUST_SRC_PATH` and `RUST_SRC_PATH` environment variables. This make the
behavior of racer the same in Emacs and on the command line with these variables
set. It will also allow `RUST_SRC_PATH` to be set to an empty string, which
is required with racer installed via nixpkgs, where `RUST_SRC_PATH` should not be set.

https://github.com/NixOS/nixpkgs/blob/348f938c552dcfdc41e543dd20511b5d0e4e70f6/pkgs/development/tools/rust/racer/rust-src.patch